--- mrh/__init__.py
@@ -1,2 +1,0 @@
-# flake8: noqa
-from mrh.logger import get_logger
